pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "swim-platform"
version = "1.0.0"
description = "Surface Water Intelligence & Monitoring — multi-agent HAB prediction platform"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "langchain>=0.3",
    "langgraph>=0.6",
    "langchain-google-genai>=2.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "streamlit>=1.30",
    "plotly>=5.18",
    "pydeck>=0.8",
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "httpx>=0.27",
    "pydantic>=2.5",
    "python-dotenv>=1.0",
    "pyarrow>=14.0",
    "pillow>=10.0",
    "geopy>=2.4",
    "shapely>=2.0",
    "a2a-sdk[http-server]",
    "prometheus_client>=0.20",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21", "ruff>=0.1"]

[project.scripts]
swim-launch = "swim.launcher:main"
swim-main = "swim.agents.main_agent.controller:main"
swim-api = "swim.api.endpoints:main"
swim-retrain = "swim.agents.predikt.retraining:retrain"

[tool.setuptools.packages.find]
where = ["."]
include = ["swim*"]
